+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedImageDataLoader.swift
@@ -1,0 +1,57 @@
+//
+//  LocalFeedImageDataLoader.swift
+//  EssentialFeed
+//
+//  Created by Anthony on 13/4/25.
+//
+
+import Foundation
+
+public final class LocalFeedImageDataLoader: FeedImageDataLoader {
+    let store: FeedImageDataStore
+    
+    public init(store: FeedImageDataStore) {
+        self.store = store
+    }
+    
+    public enum Error: Swift.Error {
+        case failed
+        case notFound
+    }
+    
+    class Task: FeedImageDataLoaderTask {
+        var completion: ((FeedImageDataLoader.Result) -> Void)?
+        
+        init(completion: @escaping ((FeedImageDataLoader.Result) -> Void)) {
+            self.completion = completion
+        }
+        
+        func complete(with result: FeedImageDataLoader.Result) {
+            completion?(result)
+        }
+        
+        func cancel() {
+            preventFutherCompletions()
+        }
+        
+        private func preventFutherCompletions() {
+            completion = nil
+        }
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        
+        let task = Task(completion: completion)
+        store.retrieve(dataForURL: url) { [weak self] result in
+            guard self != nil else { return }
+            
+            task.complete(with: result
+                .mapError { _ in LocalFeedImageDataLoader.Error.failed }
+                .flatMap { data in
+                    data.map { .success($0) } ?? .failure(LocalFeedImageDataLoader.Error.notFound)
+                })
+        }
+        
+        return task
+    }
+}
